Fix circular import issue by moving _get_target_group_name function to cdk_builder_util.py
